fixing jobcontrol

+++ includes/struct.h
@@ -6,7 +6,7 @@
 /*   By: arsciand <arsciand@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/02/06 21:52:21 by arsciand          #+#    #+#             */
-/*   Updated: 2020/03/02 15:24:50 by arsciand         ###   ########.fr       */
+/*   Updated: 2020/03/12 16:40:40 by mpivet-p         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -164,7 +164,7 @@ typedef struct				s_process
 	pid_t					pgid;
 	int8_t					completed;
 	int8_t					stopped;
-	int8_t					status;
+	int32_t					status;
 	int8_t					blt;
 	int						pipe[2];
 	int						close[2];
